+++ packages/shards-meteor-accounts-antd/src/client/ui/ForgotPasswordForm.tsx
@@ -1,10 +1,10 @@
-import React, { useState } from 'react';
-import {
-    ForgotPasswordFormUI,
-    ForgotPasswordFormUIProps,
-} from '/packages/shards-accounts-ui-antd/src/ForgotPasswordFormUI';
+import * as React from 'react';
+import { useState } from 'react';
+import { ForgotPasswordFormUI, ForgotPasswordFormUIProps } from '@moxb/shards-account-ui-antd';
 
-interface ForgotPasswordFormProps extends Pick<ForgotPasswordFormUIProps, 'splash'> {}
+type ForgotPasswordFormProps = Pick<ForgotPasswordFormUIProps, 'splash'>;
+
+import { loginLink, registerLink } from './links';
 
 export function ForgotPasswordForm(props: ForgotPasswordFormProps) {
     const [error, setError] = useState<string | undefined>();
@@ -17,14 +17,16 @@ export function ForgotPasswordForm(props: ForgotPasswordFormProps) {
             error={error}
             message={message}
             resetPending={resetPending}
+            loginLink={loginLink}
+            registerLink={registerLink}
             onReset={({ username }) => {
                 setResetPending(true);
                 setError(undefined);
                 // console.log('Attempting to reset', username);
-                Accounts.forgotPassword({ email: username }, (error: any) => {
+                Accounts.forgotPassword({ email: username }, (forgotError: any) => {
                     setResetPending(false);
-                    if (error) {
-                        setError(error.reason || 'Failed to initiate password reset');
+                    if (forgotError) {
+                        setError(forgotError.reason || 'Failed to initiate password reset');
                     } else {
                         setMessage('Email sent. Please check your email.');
                     }