+++ packages/shards-meteor-accounts-antd/src/index.ts
@@ -1,0 +1,3 @@
+export { LogoutButton } from './client/ui/LogoutButton';
+export { LoginRequired } from './client/routing/LoginRequired';
+export { setupUserAccounts } from './server/index';